Extract remove_method_expr_for into the shared lib crate (BT-3189)
The test-harness "mirror" of this parsing logic in
tests/repl_protocol.rs had drifted from the real implementation twice
in a row despite a "Mirrors ..." comment, since commands/repl/mod.rs
is private to the binary target and the integration test can't call
it directly. Move the function into a new beamtalk_cli::repl_meta_exprs
lib module (the established pattern this crate already uses for
sharing code between the binary and its tests/other crates — see
erlc, native_type_specs, etc.) so both the real REPL dispatch and the
integration test call the identical implementation. Structural drift
between the two is now impossible rather than merely avoided by
convention. BT-3196 tracks doing the same for :flush <sel>'s
equivalent duplication.

Also add the same not is_pid(pid) guard remove_method/5's sibling
save_definition/3 already has, for consistency with every other
Facade.dispatch(:eval, ...) call site in this file — latent today
(the cockpit UI is gated behind @connected), but matches the file's
own defensive convention should that invariant ever change.
