+++ crates/beamtalk-cli/src/lib.rs
@@ -14,6 +14,8 @@
 //!   `beamtalk-mcp` (BT-2858)
 //! - [`pid_liveness`] — cross-platform "is this PID alive?" check, shared
 //!   with `tests/cli_common` (BT-3077)
+//! - [`repl_meta_exprs`] — REPL meta-command → Beamtalk-expression builders,
+//!   shared with `tests/repl_protocol.rs` (BT-3189)
 
 pub mod build_layout;
 pub mod dependency_classes;
@@ -21,4 +23,5 @@ pub mod erlc;
 pub mod manifest;
 pub mod native_type_specs;
 pub mod pid_liveness;
+pub mod repl_meta_exprs;
 pub mod repl_startup;
